refactored dialogStyles file to main styles folder
