+++ grantly/src/components/grants/dialogs/SuggestionDialog.js
@@ -11,7 +11,7 @@ import {
   Grid,
   Typography
 } from "@material-ui/core";
-import { dialogStyles } from "../styles/DialogStyles";
+import { dialogStyles } from "../../../styles/dialogStyles"
 import { submitSuggestion } from "../../../actions/index";
 
 // When user clicks on
