+++ grantly/src/styles/dialogStyles.js
@@ -1,8 +1,5 @@
 import { makeStyles } from "@material-ui/core/styles";
 
-// We have styles in two different main places, could be combined or cleaned up
-// Using the makeStyles function to apply CSS in JS styles
-
 export const dialogStyles = makeStyles(theme => ({
   applyButton: {
     color: "#fff",
@@ -34,6 +31,15 @@ export const dialogStyles = makeStyles(theme => ({
     // paddingBottom: 20,
     paddingRight: 50,
     paddingLeft: 50
+  },
+  appBar: {
+    position: "relative",
+    // marginBottom: "5rem"
+    color: "white"
+  },
+  title: {
+    marginLeft: theme.spacing(2),
+    flex: 1
   }
 }));
